se agrega control de error para no atenticación por token

+++ AutoElite/src/main/java/com/autoelite/AutoElite/Calificacion/CalificacionController.java
@@ -16,7 +16,7 @@ public class CalificacionController {
     }
 
     @GetMapping("{idReceiver}")
-    public ResponseEntity<?> getAllCalificacion(@PathVariable("idReceiver") Integer id) {
+    public ResponseEntity<?> getAllCalificacion(@PathVariable("idReceiver") Long id) {
         Usuario receiver = new Usuario();
         receiver.setId(id);
         return calificacionService.getAllCalificacion(receiver);
